fix: move PackageInfo interfaces into blueprints-integration
