+++ meteor/lib/collections/MediaObjects.ts
@@ -3,6 +3,7 @@ import { registerCollection, ProtectedString } from '../lib'
 import { createMongoCollection } from './lib'
 import { StudioId } from './Studios'
 import { registerIndex } from '../database'
+import { PackageInfo } from '@sofie-automation/blueprints-integration'
 
 /** A string, identifying a MediaObj */
 export type MediaObjId = ProtectedString<'MediaObjId'>
@@ -88,31 +89,18 @@ export interface MediaFormat {
 	max_bit_rate?: number
 }
 
-export enum FieldOrder {
-	Unknown = 'unknown',
-	Progressive = 'progressive',
-	TFF = 'tff',
-	BFF = 'bff',
-}
-
 export interface Metadata {
 	scenes?: Array<number>
-	blacks?: Array<Anomaly>
-	freezes?: Array<Anomaly>
+	blacks?: Array<PackageInfo.Anomaly>
+	freezes?: Array<PackageInfo.Anomaly>
 }
 
 export interface MediaInfo extends Metadata {
 	name: string
-	field_order?: FieldOrder
+	field_order?: PackageInfo.FieldOrder
 	streams?: MediaStream[]
 	format?: MediaFormat
 	timebase?: number
-}
-
-export interface Anomaly {
-	start: number
-	duration: number
-	end: number
 }
 
 export interface MediaAttachment {
