+++ packages/blueprints-integration/src/index.ts
@@ -9,6 +9,7 @@ export * from './lib'
 export * from './message'
 export * from './migrations'
 export * from './package'
+export * from './packageInfo'
 export * from './rundown'
 export * from './showStyle'
 export * from './studio'
